Changed wrong data type.

+++ app/src/main/java/io/github/dnloop/noteapp/auxiliary/Validator.kt
@@ -2,6 +2,6 @@ package io.github.dnloop.noteapp.auxiliary
 
 class Validator {
     companion object {
-        fun isCategoryNull(item: Int?) = item?.toString() ?: ""
+        fun isCategoryNull(item: Long?) = item?.toString() ?: ""
     }
 }
